+++ src/main/java/com/comeon/backend/jwt/application/TokenType.java
@@ -1,4 +1,4 @@
-package com.comeon.backend.auth.application;
+package com.comeon.backend.jwt.application;
 
 import java.util.Arrays;
