+++ src/main/java/com/comeon/backend/jwt/domain/JwtBuilder.java
@@ -1,4 +1,4 @@
-package com.comeon.backend.auth.domain;
+package com.comeon.backend.jwt.domain;
 
 public interface JwtBuilder {
