+++ src/main/java/com/comeon/backend/jwt/domain/RefreshTokenRepository.java
@@ -1,4 +1,4 @@
-package com.comeon.backend.auth.domain;
+package com.comeon.backend.jwt.domain;
 
 import java.time.Instant;
 import java.util.Optional;
