+++ src/main/java/com/comeon/backend/jwt/infra/JwtBuilderImpl.java
@@ -1,7 +1,7 @@
-package com.comeon.backend.auth.infra;
+package com.comeon.backend.jwt.infra;
 
-import com.comeon.backend.auth.application.TokenType;
-import com.comeon.backend.auth.domain.JwtBuilder;
+import com.comeon.backend.jwt.application.TokenType;
+import com.comeon.backend.jwt.domain.JwtBuilder;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.security.Keys;
